DELETE METHOD ERROR CONDITION

+++ Week4/Day4/api-demo/youtuber-demo.js
@@ -71,11 +71,18 @@ app.post('/youtubers', function (req, res) {
 app.delete('/youtubers/:id', function (req, res) {
     let { id } = req.params;
     id = parseInt(id);
+    let youtuber = db.get(id);
 
-    const channelTitle = db.get(id).channelTitle;
-     db.delete(id);
-    res.json({
-        message:`${channelTitle}님 다음에 또 뵙겠습니다.`
-    })
+    if (youtuber == undefined) {
+        res.json({
+            message: `요청하신 ${id}번은 없는 정보입니다`
+        })
+    } else {
+        const channelTitle =youtuber.channelTitle;
+        db.delete(id);
+        res.json({
+            message: `${channelTitle}님 다음에 또 뵙겠습니다.`
+        })
+    }
     
 })
